+++ dashboard/functions/electricity_mix/Get_WorldElecMix.py
@@ -25,16 +25,22 @@ df["Total electricity generation"] = (
 )
 
 # Sélection des codes ISO sans aucunes données (intensité = moyenne mondiale)
-default_ISO = df[df['Year'].isna()].iloc[:,:4]
-default_ISO = default_ISO.fillna('/')
+default_ISO = df[df["Year"].isna()].iloc[:, :4]
+default_ISO = default_ISO.fillna("/")
 
 # Sélection des dernières données disponibles pour les autres pays (last measure)
 
-df = df.dropna(how='any')
-measure = df.groupby('ISO',as_index = False).agg({'Year':'max'}).rename(columns={"Year": "Last Year"})
+df = df.dropna(how="any")
+measure = (
+    df.groupby("ISO", as_index=False)
+    .agg({"Year": "max"})
+    .rename(columns={"Year": "Last Year"})
+)
 
 select = pd.merge(df, measure, on=["ISO"], how="inner")
-select = select[select['Year']==select['Last Year']].drop('Last Year', axis=1) #.rename(columns={"Year": "Last measure"})
+select = select[select["Year"] == select["Last Year"]].drop(
+    "Last Year", axis=1
+)  # .rename(columns={"Year": "Last measure"})
 
 # Préparation du fichier final pour le Mix électrique mondial
 
@@ -46,13 +52,13 @@ mix["% Nuclear"] = round(select.iloc[:, 7] / select.iloc[:, 10] * 100, 1)
 mix["% Solar"] = round(select.iloc[:, 8] / select.iloc[:, 10] * 100, 1)
 mix["% Wind"] = round(select.iloc[:, 9] / select.iloc[:, 10] * 100, 1)
 
-default_ISO["% Fossil"] = '/'
-default_ISO["% Geothermal"] = '/'
-default_ISO["% Hydro"] = '/'
-default_ISO["% Nuclear"] = '/'
-default_ISO["% Solar"] = '/'
-default_ISO["% Wind"] = '/'
+default_ISO["% Fossil"] = "/"
+default_ISO["% Geothermal"] = "/"
+default_ISO["% Hydro"] = "/"
+default_ISO["% Nuclear"] = "/"
+default_ISO["% Solar"] = "/"
+default_ISO["% Wind"] = "/"
 
-mix = pd.concat([mix,default_ISO])
+mix = pd.concat([mix, default_ISO])
 
 mix.to_csv("dashboard/WorldElectricityMix.csv", index=False)
