+++ dashboard/functions/electricity_mix/WorldIntensity.py
@@ -5,6 +5,7 @@
 # sys.path.append(os.path.abspath(currentdir))
 
 import pandas as pd
+
 from codecarbon.core.emissions import Emissions
 from codecarbon.core.units import Energy
 from codecarbon.external.geography import GeoMetadata
