+++ dashboard/requirements-new.txt
@@ -1,0 +1,4 @@
+codecarbon>=2.0.0
+dash>=2.2.0
+plotly>=5.6.0
+dash_bootstrap_components
