+++ dashboard/requirements.txt
@@ -1,3 +1,3 @@
-codecarbon==2.0.0a1
-dash==2.2.0
-plotly==5.6.0
+codecarbon>=2.0.0
+dash>=2.2.0
+plotly>=5.6.0
